src/pyproject.toml
[project]
name = "gugutos_player"
version = "0.0.1"

dependencies = [
	"spotipy >= 2.23.0, < 3.0.0",
	"pyserial >=3.5, < 4.0",
	"RPi.GPIO >= 0.7.1, < 1.0.0",
	"PyYAML >= 6.0, < 7.0"
]

[build-system]
requires = [
    "setuptools >= 67.8.0",
    "setuptools_scm >= 7.1.0"
]